+++ src/components/FrontPage/StartButton.scss
@@ -1,0 +1,4 @@
+.center {
+  display: flex;
+  justify-content: center;
+}
