--- src/components/FrontPage/WelcomeBox.scss
@@ -1,36 +1,0 @@
-@import '../../styles/variables';
-
-.center {
-  display: flex;
-  justify-content: center;
-}
-
-.infoBox {
-  padding: 10px;
-  max-width: 700px;
-  background-color: $color-green-lighter;
-  border: 1px solid #b2b2b2;
-  border-radius: 10px;
-  font-family: $font-family-roboto;
-  font-size: 16px;
-  font-weight: 300;
-  position: relative;
-}
-
-.xSign {
-  &,
-  &:hover,
-  &:active:focus,
-  &:focus {
-    background: transparent;
-    text-decoration: none;
-    border: none;
-    outline: none;
-  }
-  color: gray;
-  font-size: 1.1em;
-  transition: 0.1s;
-  position: absolute;
-  right: 5px;
-  top: 5px;
-}
