+++ src/constants/captions_en.js
@@ -16,15 +16,8 @@ export default {
   frontpage: {
     courses: 'Courses',
     otherwebsitecourses: 'Courses on other webpages',
-    welcomebox: {
-      header: 'Hello! You are now in studentmode',
-      changemode: 'Are you not a student? Click the student / teacher button in the navigation menu to switch mode. ' +
-                  'You can also choose to hide this box forever, by pressing X in the corner',
-      info: 'Welcome to Kodeklubbens lesson pages! Here you will find many guides that you can use as inspiration ' +
-            'to teach you programming and create your own games, apps and web pages. For beginners, we recommend ' +
-            'taking a look at the block-based tasks in Code Studio or Scratch. Please use the filters on the left ' +
-            'side to find tasks that are right for you!',
-      startbutton: 'Start here!',
+    startbutton: {
+      begin: 'Start here!',
       buttonlink: '/scratch/astrokatt/astrokatt',
       continuebutton: 'Continue...',
     },
